The AffinityWheel class got a change to be a bit more dynamic in text positioning and size changes when drawing the affinity triangle.
The PineSaveGameEditor class saw the biggest set of changes that actually allow the user to modify item values, affinity values, and then save the changes back to the file that is currently selected. There were some UI changes as well to remove items that were no longer used or necessary.